Add files via upload
In this project, each of the Convolution, Maxpooling, FullyConnected, Activation (Relu, Tanh, Sigmoid), Softmax, Linear and Forward and Backward functions are written, and then using these layers, the Lenet-5 network which It includes three Convolution layers, two Maxpooling, FullyConnected and three Relu, and Train and Test of this network is implemented using Dataset Mnist. At the end of the Test section, the accuracy of the network is printed, and Matplotlib is also used to display the input Image in this section.